genX_mfd: remove one-to-one mapping for picture_type
In genX_mfd_vc1_pic_state, the logic of picture_type is not very clear
as it is mapped from the input range to the hardware range. The mapping,
however, is an identity mapping that, effectively, only maps a skipped
picture to a P picture. By putting the skipped picture logic explicitly
in the ocde, we can remove the unnecessary mapping and make the logic
much more clear.

Signed-off-by: Jerome Borsboom <jerome.borsboom@carpalis.nl>